requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
uvicorn==0.35.0
